+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Extensions
 - `UIColor.dynamic(light:dark:)` method is added, that generates it's color data dynamically
 - `NSColor.dynamic(light:dark:)` method is added, that generates it's color data dynamically
 - `View.onFirstDisappear(perform:)` method is added, that performs an action before `View` disappears for the first time
+- `AttributedString.init(_:attributeContainers:)` initializer is added, that initializes `AttributedString` with child `AttributedString` components created from mapping tag names to `AttributeContainer`s
 
 ### [5.3.1(90)](https://github.com/VakhoKontridze/VCore/releases/tag/5.3.1) — *2024 01 23*